requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
asyncpg
databases
python-dotenv
pyjwt[crypto]
grpcio
grpcio-tools
pytest
httpx
asgi-lifespan
psycopg2
pydantic[email]
pytest-asyncio
slowapi
grpcio-reflection
python-multipart
psutil
